requirements.txt
python-nmap
scapy
rich